requirements.txt
dashscope>=1.18.0
pdfplumber>=0.11.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0